+++ examples/NOC/ch02/forces_many_2_2.clj
@@ -1,4 +1,4 @@
-(ns examples.ex34-noc22-forces-many
+(ns examples.NOC.ch02.forces-many-2-2
   (:require [clojure2d.core :refer :all]
             [clojure2d.math.vector :as v]
             [clojure2d.math :as m]
@@ -11,19 +11,19 @@
 (def ^:const ^int w 640)
 (def ^:const ^int h 360)
 
-(def ^:const ^int number-of-movers 30)
+(def ^:const ^int number-of-movers 20)
 
 (def wind (Vec2. 0.01 0.0))
 (def gravity (Vec2. 0.0 0.1))
 
-(defrecord Mover [position velocity mass])
+(deftype Mover [position velocity ^double mass])
 
 (defn make-mover
   "Create Mover"
   []
-  (->Mover (Vec2. 30.0 30.0)
-           (Vec2. (r/drand -1.0 1.0) (r/drand -1.0 1.0))
-           (r/drand 0.2 2.0)))
+  (->Mover (Vec2. 0.0 0.0)
+           (Vec2. 0.0 0.0)
+           (r/drand 0.1 4.0)))
 
 (defn apply-force
   "Apply force"
@@ -41,29 +41,29 @@
 
 (defn move-mover
   "Move mover"
-  [m]
+  [^Mover m]
   (let [acc (-> (Vec2. 0.0 0.0)
-                (apply-force wind (:mass m))
-                (apply-force gravity (:mass m)))
-        vel (v/add (:velocity m) acc)
-        pos (v/add (:position m) vel)
+                (apply-force wind (.mass m))
+                (apply-force gravity (.mass m)))
+        vel (v/add (.velocity m) acc)
+        pos (v/add (.position m) vel)
         [new-vel new-pos] (check-edges vel pos)]
-    (->Mover new-pos new-vel (:mass m))))
+    (->Mover new-pos new-vel (.mass m))))
 
 (defn draw-and-move
   "Draw mover, move and return new one."
-  [canvas m]
-  (let [size (* 48.0 ^double (:mass m))]
-    (set-color canvas 127 127 127 127)
-    (ellipse canvas (.x ^Vec2 (:position m)) (.y ^Vec2 (:position m)) size size false)
-    (set-stroke canvas 1.2)
+  [canvas ^Mover m]
+  (let [size (* 16.0 ^double (.mass m))]
+    (set-color canvas 0 0 0 127)
+    (ellipse canvas (.x ^Vec2 (.position m)) (.y ^Vec2 (.position m)) size size false)
+    (set-stroke canvas 2)
     (set-color canvas 0 0 0)
-    (ellipse canvas (.x ^Vec2 (:position m)) (.y ^Vec2 (:position m)) size size true)
+    (ellipse canvas (.x ^Vec2 (.position m)) (.y ^Vec2 (.position m)) size size true)
     (move-mover m)))
 
 (defn draw
   "Draw movers on canvas"
-  [canvas _ last-m]
+  [canvas _ _ last-m]
   (set-background canvas 255 255 255)
   (let [m (or last-m (repeatedly number-of-movers make-mover))]
     (mapv (partial draw-and-move canvas) m)))